requirements/base.txt
Django
argparse
django-extensions
ipython
psycopg2
six
sqlparse
git+https://github.com/scotm/django-generic-json-views.git
Unipath
factory-boy
django-sortedm2m
jsonfield
django-braces
django-filter
django-htmlmin
django-flat-theme
django-memoize

